Match navbar styles with wireframe
