--- src/pages/Navbar.css
@@ -1,39 +1,0 @@
-.navbar-container {
-  padding: 0 4%;
-  background-color: #fff;
-  border-bottom: 1px solid #e8e8e8;
-  display: flex;
-  align-items: center;
-}
-
-.navbar-left-align {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.navbar-logo-link {
-  text-decoration: none;
-  font-size: 38px;
-  font-weight: bold;
-  color: #5b5a5a;
-}
-
-.navbar-ul {
-  display: flex;
-  list-style-type: none;
-  align-items: center;
-  justify-content: center;
-}
-
-.navbar-li-item {
-  text-decoration: none;
-  color: rgb(13, 13, 13);
-  margin-right: 20px;
-}
-
-.active {
-  text-decoration: underline;
-  font-weight: bold;
-  color: #1e4fb1;
-}
